+++ libexec/scoop-alias.ps1
@@ -39,7 +39,7 @@ if ($SubCommand -notin $SubCommands) {
     exit 1
 }
 
-$opt, $other, $err = getopt $Args 'v' , 'verbose'
+$opt, $other, $err = getopt $Args 'v' 'verbose'
 if ($err) { "scoop alias: $err"; exit 1 }
 
 $name, $command, $description = $other